trait CanOpenUrl trait
