+++ packages/forms/src/Components/Select.php
@@ -11,6 +11,7 @@ class Select extends Field
     use Concerns\HasOptions;
     use Concerns\HasPlaceholder;
     use Concerns\HasAffixes;
+    use Concerns\CanOpenUrl;
 
     protected string $view = 'forms::components.select';
 
@@ -29,8 +30,6 @@ class Select extends Field
     protected string | Closure | null $noSearchResultsMessage = null;
 
     protected string | Closure | null $searchPrompt = null;
-
-    protected string | Closure | null $url = null;
 
     protected function setUp(): void
     {
@@ -116,13 +115,6 @@ class Select extends Field
         return $this;
     }
 
-    public function url(string | Closure | null $url): static
-    {
-        $this->url = $url;
-
-        return $this;
-    }
-
     public function getOptionLabel(): ?string
     {
         return $this->evaluate($this->getOptionLabelUsing, [
@@ -160,11 +152,6 @@ class Select extends Field
         }
 
         return $results;
-    }
-
-    public function getUrl(): ?string
-    {
-        return $this->evaluate($this->url);
     }
 
     public function isOptionDisabled($value, string $label): bool
